requirements.txt
google-generativeai
google-ai-generativelanguage
sentence-transformers
langchain-community
langchain
nltk
streamlit